requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
aiohttp>=3.8.5
openai>=1.3.7
anthropic>=0.5.0
pandas>=2.1.1
numpy>=1.24.3
aiofiles>=23.2.1
python-multipart>=0.0.6